% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_answer.R
\name{retrieve_answer}
\alias{retrieve_answer}
\title{Function to retrieve the answers of the data transformation
This function is used to indicate a desired option and retrieve the answer}
\usage{
retrieve_answer(opcion)
}
\arguments{
\item{opcion}{Number (Integer, between 1 and 6)}
}
\value{
Displays the selected answer on the console
}
\description{
Function to retrieve the answers of the data transformation
This function is used to indicate a desired option and retrieve the answer
}
\examples{
retrieve_answer (2)
retrieve_answer (5)

}
